+++ src/application/mod.rs
@@ -22,6 +22,7 @@ pub mod forget;
 pub mod from_message;
 pub mod item_fields;
 pub mod local_folders;
+pub mod long_text;
 pub mod mail_auth;
 pub mod mail_controller;
 pub mod mail_sync;